Google Chome OS Patch: update-type-ownerauth.patch
Adds a new "tpm12-ownerauth" update type. This can be used on TPMs that already
have an owner. The owner password is assumed to be the well-known secret, i.e.
the auth secret used for communication with the TPM is the SHA-1 hash of 20
zero bytes. We check that owner authentication using this secret is successful
during the preparation phase before the actual update attempt starts.